Add new file for header
- Move stylesheets to new folder
