+++ src/assets/stylesheets/_mixins.scss
@@ -6,7 +6,7 @@
 }
 
 @mixin background-image($url, $position, $repeat, $size) {
-  background-image: url('./assets/images/#{$url}');
+  background-image: url('../images/#{$url}');
   background-position: $position;
   background-repeat: $repeat;
   background-size: $size;
